Move texture code to a separate file
This is going to grow once we add support for more glTex* functions.
